Split read unit tests
